pyproject.toml
[project]
name = "langgraph-learn"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.14"
dependencies = [
    "langchain>=1.2.9",
    "langchain-anthropic>=1.3.2",
    "langchain-google-genai>=4.2.0",
    "langchain-openai>=1.1.7",
    "langgraph>=1.0.8",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
]
